requirements.txt
beautifulsoup4~=4.9.3
certifi~=2020.6.20
chardet~=3.0.4
cobble~=0.1.3
idna~=2.10
lxml~=4.6.1
mammoth~=1.4.13
python-docx~=0.8.10
requests~=2.24.0
selenium~=3.141.0
soupsieve~=2.0.1
urllib3~=1.25.11